Fixed image barriers to apply to arraySlice
These barriers on the destination side were always on slice 0. For the "Subimage Tests" case, the test writes static data to both slice 0 and slice 1. When we wrote them to slice 1, it transitioned the layouts on since 0 instead. Drivers are allowed to clear in this case, and my particular driver happily did.