Action test fixes
* Fixed test and check happening from two different input frames
The user is prompted to get a float input to a particular value. xrSyncActions is called repeatedly to test to see if that has happened.  Then xrSyncActions is called one last time to see if the float value was coerced to the correct boolean value in another action. For very unstable float values (like x=1 on a capsule-shaped trackpad) or slow-running implementation the float value could have changed between those two frames, and the test would fail.

This change piggybacks the bool=Y test into the same xrSyncActions call as the float=X test

* Fixed X and Y times not exactly matching XY time

The spec guarantees that action update times not change until the next xrSyncActions call, but doesn't require that the X and Y components have the same timestamp as the combined vector XY time. In SteamVR these components are updated asynchronously by drivers, so they may not match.

The test will now make sure the XY time is the most recent of its two components.