Changes to make some kinds of layer comp flag failures more obvious
The background alpha is on the green layers and the example quad are 0 now. Those quad layers don't include the "use texture alpha" flag, so they should draw in exactly the same way. If the implementation is actually using the alpha from the texture, they will be invisible, which will be obviously wrong in the output.

Added the "not premult" flag to those quads too. SteamVR doesn't currently pay any attention to the "use alpha" flag, but it does monitor the premult flag.  A fully conformant runtime should ignore the premult flag if the other one isn't set too, so this should not affect those.